Run travis-after-success script only once
This script contains various actions to be done after tests succeeds -
it is wasteful or even wrong to run it for every matrix row.

+++ ci/travis-after-success.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 set -xe
 
+test "$AFTER_SCRIPT" = "1" || exit 0
+
 # Submit code coverage report
 docker run -ti --rm -v $PWD:/docker -w /docker \
     -e CI \
